+++ src/game/Object/Player.cpp
@@ -5116,104 +5116,6 @@ void Player::SendExplorationExperience(uint32 Area, uint32 Experience)
     GetSession()->SendPacket(&data);
 }
 
-void Player::SendDungeonDifficulty(bool IsInGroup)
-{
-    uint8 val = 0x00000001;
-    WorldPacket data(MSG_SET_DUNGEON_DIFFICULTY, 12);
-    data << (uint32)GetDifficulty();
-    data << uint32(val);
-    data << uint32(IsInGroup);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends a reset-failed notification for an instance map.
- *
- * @param mapid The map identifier that failed to reset.
- */
-void Player::SendResetFailedNotify(uint32 mapid)
-{
-    WorldPacket data(SMSG_RESET_FAILED_NOTIFY, 4);
-    data << uint32(mapid);
-    GetSession()->SendPacket(&data);
-}
-
-/// Reset all solo instances and optionally send a message on success for each
-void Player::ResetInstances(InstanceResetMethod method)
-{
-    // method can be INSTANCE_RESET_ALL, INSTANCE_RESET_CHANGE_DIFFICULTY, INSTANCE_RESET_GROUP_JOIN
-
-    // we assume that when the difficulty changes, all instances that can be reset will be
-    Difficulty diff = GetDifficulty();
-
-    for (BoundInstancesMap::iterator itr = m_boundInstances[diff].begin(); itr != m_boundInstances[diff].end();)
-    {
-        DungeonPersistentState* state = itr->second.state;
-        const MapEntry* entry = sMapStore.LookupEntry(itr->first);
-        if (!entry || !state->CanReset())
-        {
-            ++itr;
-            continue;
-        }
-
-        if (method == INSTANCE_RESET_ALL)
-        {
-            // the "reset all instances" method can only reset normal maps
-            if (entry->map_type == MAP_RAID || diff == DUNGEON_DIFFICULTY_HEROIC)
-            {
-                ++itr;
-                continue;
-            }
-        }
-
-        // if the map is loaded, reset it
-        if (Map* map = sMapMgr.FindMap(state->GetMapId(), state->GetInstanceId()))
-            if (map->IsDungeon())
-            {
-                ((DungeonMap*)map)->Reset(method);
-            }
-
-        // since this is a solo instance there should not be any players inside
-        if (method == INSTANCE_RESET_ALL || method == INSTANCE_RESET_CHANGE_DIFFICULTY)
-        {
-            SendResetInstanceSuccess(state->GetMapId());
-        }
-
-        state->DeleteFromDB();
-        m_boundInstances[diff].erase(itr++);
-
-        // the following should remove the instance save from the manager and delete it as well
-        state->RemovePlayer(this);
-    }
-}
-
-/**
- * @brief Sends a successful instance reset notification to the client.
- *
- * @param MapId The reset map identifier.
- */
-void Player::SendResetInstanceSuccess(uint32 MapId)
-{
-    WorldPacket data(SMSG_INSTANCE_RESET, 4);
-    data << uint32(MapId);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends an instance reset failure message to the client.
- *
- * @param reason The reset failure reason code.
- * @param MapId The map identifier that failed to reset.
- */
-void Player::SendResetInstanceFailed(uint32 reason, uint32 MapId)
-{
-    // reason: see enum InstanceResetFailReason
-    WorldPacket data(SMSG_INSTANCE_RESET_FAILED, 4);
-    data << uint32(reason);
-    data << uint32(MapId);
-    GetSession()->SendPacket(&data);
-}
-
 /*********************************************************/
 /***              Update timers                        ***/
 /*********************************************************/
@@ -6774,40 +6676,6 @@ void Player::SendTransferAbortedByLockStatus(MapEntry const* mapEntry, AreaLockS
 }
 
 /**
- * @brief Sends an instance reset warning message to the client.
- *
- * @param mapid The map identifier being reset.
- * @param time The remaining time until reset in seconds.
- */
-void Player::SendInstanceResetWarning(uint32 mapid, uint32 time)
-{
-    // type of warning, based on the time remaining until reset
-    uint32 type;
-    if (time > 3600)
-    {
-        type = RAID_INSTANCE_WELCOME;
-    }
-    else if (time > 900 && time <= 3600)
-    {
-        type = RAID_INSTANCE_WARNING_HOURS;
-    }
-    else if (time > 300 && time <= 900)
-    {
-        type = RAID_INSTANCE_WARNING_MIN;
-    }
-    else
-    {
-        type = RAID_INSTANCE_WARNING_MIN_SOON;
-    }
-
-    WorldPacket data(SMSG_RAID_INSTANCE_MESSAGE, 4 + 4 + 4);
-    data << uint32(type);
-    data << uint32(mapid);
-    data << uint32(time);
-    GetSession()->SendPacket(&data);
-}
-
-/**
  * @brief Applies the default equip cooldown for item use spells.
  *
  * @param pItem The item whose on-use spells should receive equip cooldowns.
@@ -7377,171 +7245,6 @@ uint32 Player::GetResurrectionSpellId()
     }
 
     return spell_id;
-}
-
-// Used in triggers for check "Only to targets that grant experience or honor" req
-bool Player::isHonorOrXPTarget(Unit* pVictim) const
-{
-    uint32 v_level = pVictim->getLevel();
-    uint32 k_grey  = MaNGOS::XP::GetGrayLevel(getLevel());
-
-    // Victim level less gray level
-    if (v_level <= k_grey)
-    {
-        return false;
-    }
-
-    if (pVictim->GetTypeId() == TYPEID_UNIT)
-    {
-        if (((Creature*)pVictim)->IsTotem() ||
-            ((Creature*)pVictim)->IsPet() ||
-            ((Creature*)pVictim)->GetCreatureInfo()->ExtraFlags & CREATURE_FLAG_EXTRA_NO_XP_AT_KILL)
-            {
-                return false;
-            }
-    }
-    return true;
-}
-
-/**
- * @brief Rewards the player for killing a unit outside group reward distribution.
- *
- * @param pVictim The killed unit.
- */
-void Player::RewardSinglePlayerAtKill(Unit* pVictim)
-{
-    bool PvP = pVictim->IsCharmedOwnedByPlayerOrPlayer();
-
-    uint32 xp = PvP ? 0 : MaNGOS::XP::Gain(this, pVictim);
-
-    // honor can be in PvP and !PvP (racial leader) cases
-    RewardHonor(pVictim, 1);
-
-    // xp and reputation only in !PvP case
-    if (!PvP)
-    {
-        RewardReputation(pVictim, 1);
-        GiveXP(xp, pVictim);
-
-        if (Pet* pet = GetPet())
-        {
-            pet->GivePetXP(xp);
-        }
-
-        // normal creature (not pet/etc) can be only in !PvP case
-        if (pVictim->GetTypeId() == TYPEID_UNIT)
-        {
-            if (CreatureInfo const* normalInfo = ObjectMgr::GetCreatureTemplate(pVictim->GetEntry()))
-            {
-                KilledMonster(normalInfo, pVictim->GetObjectGuid());
-            }
-        }
-    }
-}
-
-/**
- * @brief Grants event kill credit to the player or nearby group members.
- *
- * @param creature_id The credited creature entry identifier.
- * @param pRewardSource The world object used for distance checks.
- */
-void Player::RewardPlayerAndGroupAtEvent(uint32 creature_id, WorldObject* pRewardSource)
-{
-    MANGOS_ASSERT((!GetGroup() || pRewardSource));              // Player::RewardPlayerAndGroupAtEvent called for Group-Case but no source for range searching provided
-
-    ObjectGuid creature_guid = pRewardSource && pRewardSource->GetTypeId() == TYPEID_UNIT ? pRewardSource->GetObjectGuid() : ObjectGuid();
-
-    // prepare data for near group iteration
-    if (Group* pGroup = GetGroup())
-    {
-        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* pGroupGuy = itr->getSource();
-            if (!pGroupGuy)
-            {
-                continue;
-            }
-
-            if (!pGroupGuy->IsAtGroupRewardDistance(pRewardSource))
-            {
-                continue; // member (alive or dead) or his corpse at req. distance
-            }
-
-            // quest objectives updated only for alive group member or dead but with not released body
-            if (pGroupGuy->IsAlive() || !pGroupGuy->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GHOST))
-            {
-                pGroupGuy->KilledMonsterCredit(creature_id, creature_guid);
-            }
-        }
-    }
-    else                                                    // if (!pGroup)
-    {
-        KilledMonsterCredit(creature_id, creature_guid);
-    }
-}
-
-/**
- * @brief Grants quest cast credit to the player or nearby group members.
- *
- * @param pRewardSource The credited creature or gameobject.
- * @param spellid The spell that granted the credit.
- */
-void Player::RewardPlayerAndGroupAtCast(WorldObject* pRewardSource, uint32 spellid)
-{
-    // prepare data for near group iteration
-    if (Group* pGroup = GetGroup())
-    {
-        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* pGroupGuy = itr->getSource();
-            if (!pGroupGuy)
-            {
-                continue;
-            }
-
-            if (!pGroupGuy->IsAtGroupRewardDistance(pRewardSource))
-            {
-                continue; // member (alive or dead) or his corpse at req. distance
-            }
-
-            // quest objectives updated only for alive group member or dead but with not released body
-            if (pGroupGuy->IsAlive() || !pGroupGuy->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GHOST))
-            {
-                pGroupGuy->CastedCreatureOrGO(pRewardSource->GetEntry(), pRewardSource->GetObjectGuid(), spellid, pGroupGuy == this);
-            }
-        }
-    }
-    else                                                    // if (!pGroup)
-    {
-        CastedCreatureOrGO(pRewardSource->GetEntry(), pRewardSource->GetObjectGuid(), spellid);
-    }
-}
-
-/**
- * @brief Checks whether the player or corpse is close enough for shared rewards.
- *
- * @param pRewardSource The source object used for the distance check.
- * @return True if the player qualifies for group reward range; otherwise, false.
- */
-bool Player::IsAtGroupRewardDistance(WorldObject const* pRewardSource) const
-{
-    if (pRewardSource->IsWithinDistInMap(this, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE)))
-    {
-        return true;
-    }
-
-    if (IsAlive())
-    {
-        return false;
-    }
-
-    Corpse* corpse = GetCorpse();
-    if (!corpse)
-    {
-        return false;
-    }
-
-    return pRewardSource->IsWithinDistInMap(corpse, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE));
 }
 
 /**
